Extract BftForksScheduleFactory from ForksSchedule
