+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/ForksSchedule.java
@@ -14,15 +14,9 @@
  */
 package org.hyperledger.besu.consensus.common;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
-import org.hyperledger.besu.config.BftConfigOptions;
-import org.hyperledger.besu.config.BftFork;
-
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.List;
 import java.util.NavigableSet;
 import java.util.Set;
 import java.util.TreeSet;
@@ -36,39 +30,10 @@ public class ForksSchedule<C> {
       new TreeSet<>(
           Comparator.comparing((Function<ForkSpec<C>, Long>) ForkSpec::getBlock).reversed());
 
-  public interface BftSpecCreator<T extends BftConfigOptions, U extends BftFork> {
-    T create(ForkSpec<T> lastSpec, U fork);
-  }
-
   @VisibleForTesting
   public ForksSchedule(final ForkSpec<C> genesisFork, final Collection<ForkSpec<C>> forks) {
     this.forks.add(genesisFork);
     this.forks.addAll(forks);
-  }
-
-  public static <T extends BftConfigOptions, U extends BftFork>
-      ForksSchedule<T> createForBftConfigOptions(
-          final T initial, final List<U> forks, final BftSpecCreator<T, U> specCreator) {
-    checkArgument(
-        forks.stream().allMatch(f -> f.getForkBlock() > 0),
-        "Transition cannot be created for genesis block");
-    checkArgument(
-        forks.stream().map(BftFork::getForkBlock).distinct().count() == forks.size(),
-        "Duplicate transitions cannot be created for the same block");
-
-    final NavigableSet<ForkSpec<T>> specs = new TreeSet<>(Comparator.comparing(ForkSpec::getBlock));
-    final ForkSpec<T> initialForkSpec = new ForkSpec<>(0, initial);
-    specs.add(initialForkSpec);
-
-    forks.stream()
-        .sorted(Comparator.comparing(BftFork::getForkBlock))
-        .forEachOrdered(
-            f -> {
-              final T spec = specCreator.create(specs.last(), f);
-              specs.add(new ForkSpec<>(f.getForkBlock(), spec));
-            });
-
-    return new ForksSchedule<>(initialForkSpec, specs.tailSet(initialForkSpec, false));
   }
 
   public ForkSpec<C> getFork(final long blockNumber) {
