+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/IbftForksSchedulesFactory.java
@@ -19,12 +19,13 @@ import org.hyperledger.besu.config.BftFork;
 import org.hyperledger.besu.config.GenesisConfigOptions;
 import org.hyperledger.besu.consensus.common.ForkSpec;
 import org.hyperledger.besu.consensus.common.ForksSchedule;
+import org.hyperledger.besu.consensus.common.bft.BftForksScheduleFactory;
 import org.hyperledger.besu.consensus.common.bft.MutableBftConfigOptions;
 
 public class IbftForksSchedulesFactory {
 
   public static ForksSchedule<BftConfigOptions> create(final GenesisConfigOptions genesisConfig) {
-    return ForksSchedule.createForBftConfigOptions(
+    return BftForksScheduleFactory.create(
         genesisConfig.getBftConfigOptions(),
         genesisConfig.getTransitions().getIbftForks(),
         IbftForksSchedulesFactory::createBftConfigOptions);
